// src/app/views/canopy-process/fabrication-maker/Model/fabricationprcPartDts.ts
export interface IfabricationPrcPartDts {
  Part: string;
  PartCode: string;
  UName: string;
  KitQty: number;
  BatchQty: number;
  StkQty: number;
  WtPerUnit: number;
  WtPerBatch: number;
  PRate: number;
  SRate: number;
  Length: number;
  Width: number;
  Thickness: number;
  LossWgt: number;
  sqft: number;
  categoryID: number;
}